Do not exit with success if input parameters were incorrect
I was running par_psql as part of a CI system and it looked like the job was a success. However, it had errored early but returned a successful exit code.

I also believe (please correct me if I'm wrong!) that par_psql won't work with `--single-transaction`, so I am detecting that parameter.